requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dateutils==0.6.12
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-logging==2.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.0
idna==2.10
packaging==20.9
proto-plus==1.14.2
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.3
